TVbo {lmerTest}	R Documentation

Sensory Assesment of B&O TVs

Description

The TVbo dataset has kindly been made available by the Danish high-end consumer electronics company Bang & Olufsen. The main purpose was to assess 12 different TV sets (products) specified by the two attributes Picture and TVset. 15 different response variables (characteristics of the product) were assessed by a trained panel with 8 assessors.

Usage

data(TVbo)

Format

Assessor

factor with 8 levels assessors.

TVset

product factor with 3 levels.

Picture

product factor with 4 levels.

In addition the following 15 numeric (response) variables are the characteristics on which the TV sets (products) are assessed:

Coloursaturation, Colourbalance, Noise, Depth, Sharpness, Lightlevel, Contrast, Sharpnessofmovement, Flickeringstationary, Flickeringmovement, Distortion, Dimglasseffect, Cutting, Flossyedges, Elasticeffect.

Examples

fm <- lmer(Coloursaturation ~ TVset + Picture + (1|Assessor:TVset) +
             (1|Assessor), data=TVbo)
ranova(fm)
anova(fm)

[Package lmerTest version 3.2-1 Index]