as_lmerModLmerTest {lmerTest}	R Documentation

Coerce lmerMod Objects to lmerModLmerTest

Description

Coercing an lme4::lmer model-object (of class 'lmerMod') to a model-object of class 'lmerModLmerTest' involves computing the covariance matrix of the variance parameters and the gradient (Jacobian) of cov(beta) with respect to the variance parameters.

Usage

as_lmerModLmerTest(model, tol = 1e-08)

Arguments

model
and lmer model-object (of class 'lmerMod') – the result of a call to lme4::lmer()

tol
tolerance for determining of eigenvalues are negative, zero or positive

Value

an object of class 'lmerModLmerTest' which sets the following slots:

vcov_varpar
the asymptotic covariance matrix of the variance parameters (theta, sigma).

Jac_list
list of Jacobian matrices; gradients of vcov(beta) with respect to the variance parameters.

vcov_beta
the asymptotic covariance matrix of the fixed-effect regression parameters (beta; vcov(beta)).

sigma
the residual standard deviation.

Author(s)

Rune Haubo B. Christensen

See Also

the class definition in lmerModLmerTest) and lmer

Examples

m <- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
bm <- as_lmerModLmerTest(m)
slotNames(bm)

[Package lmerTest version 3.2-1 Index]