lmerModLmerTest-class {lmerTest}	R Documentation

Represent Linear Mixed-Effects Models

Description

The lmerModLmerTest class extends lmerMod (which extends merMod) from the lme4-package.

Value

An object of class lmerModLmerTest with slots as in lmerMod objects (see merMod) and a few additional slots as described in the slots section.

Slots

vcov_varpar

a numeric matrix holding the asymptotic variance-covariance matrix of the variance parameters (including sigma).

Jac_list

a list of gradient matrices (Jacobians) for the gradient of the variance-covariance of beta with respect to the variance parameters, where beta are the mean-value parameters available in fixef(object).

vcov_beta

a numeric matrix holding the asymptotic variance-covariance matrix of the fixed-effect regression parameters (beta).

sigma

the residual standard deviation.

Author(s)

Rune Haubo B. Christensen

See Also

lmer and merMod

[Package lmerTest version 3.2-1 Index]