plot.step_list {lmerTest}	R Documentation

Plot LS-means for Backward Reduced Model

Description

Computes the LS-means for the final backward reduced model and passes these to plot.ls_means.

Usage

## S3 method for class 'step_list'
plot(
  x,
  y = NULL,
  which = NULL,
  pairwise = FALSE,
  mult = TRUE,
  level = 0.95,
  ddf = c("Satterthwaite", "Kenward-Roger"),
  ...
)

Arguments

x
a step_list object; the result of running step.

y
not used and ignored with a warning.

which
optional character vector naming factors for which LS-means should be plotted. If NULL (default) plots for all LS-means are generated.

pairwise
pairwise differences of LS-means?

mult
if TRUE and there is more than one term for which to plot LS-means the plots are organized in panels with facet_wrap.

level
confidence level.

ddf
denominator degree of freedom method.

...
currently not used.

Details

Error bars are confidence intervals - the default is 95 level can be changed.

Author(s)

Rune Haubo B. Christensen and Alexandra Kuznetsova

See Also

ls_means and plot.ls_means

Examples

## Not run: 
# Fit example model:
tv <- lmer(Sharpnessofmovement ~ TVset * Picture +
             (1 | Assessor:TVset) + (1 | Assessor:Picture) +
             (1 | Assessor:Picture:TVset) + (1 | Repeat) + (1 | Repeat:Picture) +
             (1 | Repeat:TVset) + (1 | Repeat:TVset:Picture) + (1 | Assessor),
           data = TVbo)

# Backward reduce the model:
(st <- step(tv)) # takes ~10 sec to run

# Pairwise comparisons of LS-means for Picture and TVset:
  plot(st, which=c("Picture", "TVset"), pairwise = TRUE)

## End(Not run)

[Package lmerTest version 3.2-1 Index]