show_tests.anova {lmerTest}	R Documentation

Show Hypothesis Tests in ANOVA Tables

Description

Extracts hypothesis matrices for terms in ANOVA tables detailing exactly which functions of the parameters are being tested in anova tables.

Usage

## S3 method for class 'anova'
show_tests(object, fractions = FALSE, names = TRUE, ...)

Arguments

object
an anova table with a "hypotheses" attribute.

fractions
display entries in the hypothesis matrices as fractions?

names
if FALSE column and row names of the hypothesis matrices are suppressed.

...
currently not used.

Value

a list of hypothesis matrices.

Author(s)

Rune Haubo B. Christensen

See Also

show_tests for ls_means objects.

Examples

# Fit basic model to the 'cake' data:
data("cake", package="lme4")
fm1 <- lmer(angle ~ recipe * temp + (1|recipe:replicate), cake)

# Type 3 anova table:
(an <- anova(fm1, type="3"))

# Display tests/hypotheses for type 1, 2, and 3 ANOVA tables:
# (and illustrate effects of 'fractions' and 'names' arguments)
show_tests(anova(fm1, type="1"))
show_tests(anova(fm1, type="2"), fractions=TRUE, names=FALSE)
show_tests(an, fractions=TRUE)

[Package lmerTest version 3.2-1 Index]